+++ contracts/burner/src/contract.rs
@@ -1,5 +1,5 @@
 use cosmwasm_std::{
-    log, Api, BankMsg, Binary, Env, Extern, HandleResponse, InitResponse, MigrateResponse, Order,
+    attr, Api, BankMsg, Binary, Env, Extern, HandleResponse, InitResponse, MigrateResponse, Order,
     Querier, StdError, StdResult, Storage,
 };
 
@@ -53,7 +53,7 @@ pub fn migrate<S: Storage, A: Api, Q: Querier>(
 
     Ok(MigrateResponse {
         messages: vec![send.into()],
-        log: vec![log("action", "burn"), log("payout", msg.payout)],
+        attributes: vec![attr("action", "burn"), attr("payout", msg.payout)],
         data: Some(data_msg.into()),
     })
 }