+++ contracts/reflect/src/contract.rs
@@ -1,5 +1,5 @@
 use cosmwasm_std::{
-    log, to_binary, Api, Binary, CosmosMsg, Env, Extern, HandleResponse, HumanAddr, InitResponse,
+    attr, to_binary, Api, Binary, CosmosMsg, Env, Extern, HandleResponse, HumanAddr, InitResponse,
     Querier, StdError, StdResult, Storage,
 };
 
@@ -47,7 +47,7 @@ pub fn try_reflect<S: Storage, A: Api, Q: Querier>(
     }
     let res = HandleResponse {
         messages: msgs,
-        log: vec![log("action", "reflect")],
+        attributes: vec![attr("action", "reflect")],
         data: None,
     };
     Ok(res)
@@ -67,7 +67,7 @@ pub fn try_change_owner<S: Storage, A: Api, Q: Querier>(
         Ok(state)
     })?;
     Ok(HandleResponse {
-        log: vec![log("action", "change_owner"), log("owner", owner)],
+        attributes: vec![attr("action", "change_owner"), attr("owner", owner)],
         ..HandleResponse::default()
     })
 }