+++ contracts/staking/src/contract.rs
@@ -1,5 +1,5 @@
 use cosmwasm_std::{
-    coin, log, to_binary, Api, BankMsg, Binary, Decimal, Env, Extern, HandleResponse, HumanAddr,
+    attr, coin, to_binary, Api, BankMsg, Binary, Decimal, Env, Extern, HandleResponse, HumanAddr,
     InitResponse, Querier, StakingMsg, StdError, StdResult, Storage, Uint128, WasmMsg,
 };
 
@@ -86,11 +86,11 @@ pub fn transfer<S: Storage, A: Api, Q: Querier>(
 
     let res = HandleResponse {
         messages: vec![],
-        log: vec![
-            log("action", "transfer"),
-            log("from", env.message.sender),
-            log("to", recipient),
-            log("amount", send),
+        attributes: vec![
+            attr("action", "transfer"),
+            attr("from", env.message.sender),
+            attr("to", recipient),
+            attr("amount", send),
         ],
         data: None,
     };
@@ -174,11 +174,11 @@ pub fn bond<S: Storage, A: Api, Q: Querier>(
             amount: payment.clone(),
         }
         .into()],
-        log: vec![
-            log("action", "bond"),
-            log("from", env.message.sender),
-            log("bonded", payment.amount),
-            log("minted", to_mint),
+        attributes: vec![
+            attr("action", "bond"),
+            attr("from", env.message.sender),
+            attr("bonded", payment.amount),
+            attr("minted", to_mint),
         ],
         data: None,
     };
@@ -243,11 +243,11 @@ pub fn unbond<S: Storage, A: Api, Q: Querier>(
             amount: coin(unbond.u128(), &invest.bond_denom),
         }
         .into()],
-        log: vec![
-            log("action", "unbond"),
-            log("to", env.message.sender),
-            log("unbonded", unbond),
-            log("burnt", amount),
+        attributes: vec![
+            attr("action", "unbond"),
+            attr("to", env.message.sender),
+            attr("unbonded", unbond),
+            attr("burnt", amount),
         ],
         data: None,
     };
@@ -293,10 +293,10 @@ pub fn claim<S: Storage, A: Api, Q: Querier>(
             amount: vec![balance],
         }
         .into()],
-        log: vec![
-            log("action", "claim"),
-            log("from", env.message.sender),
-            log("amount", to_send),
+        attributes: vec![
+            attr("action", "claim"),
+            attr("from", env.message.sender),
+            attr("amount", to_send),
         ],
         data: None,
     };
@@ -329,7 +329,7 @@ pub fn reinvest<S: Storage, A: Api, Q: Querier>(
             }
             .into(),
         ],
-        log: vec![],
+        attributes: vec![],
         data: None,
     };
     Ok(res)
@@ -372,7 +372,7 @@ pub fn _bond_all_tokens<S: Storage, A: Api, Q: Querier>(
             amount: balance.clone(),
         }
         .into()],
-        log: vec![log("action", "reinvest"), log("bonded", balance.amount)],
+        attributes: vec![attr("action", "reinvest"), attr("bonded", balance.amount)],
         data: None,
     };
     Ok(res)